Source tree reorganization, splitting demos apart
