+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "cosmian_js_lib",
-    "version": "1.0.4",
+    "version": "1.0.5",
     "license": "MIT",
     "description": "Cosmian javascript client library",
     "main": "dist/lib.js",
