+++ src/crypto/abe/hybrid_crypto/cover_crypt/encryption.ts
@@ -1,8 +1,9 @@
 import { webassembly_encrypt_hybrid_block, webassembly_encrypt_hybrid_header } from "cover_crypt";
 import { logger } from "../../../../utils/logger";
 import { hexEncode } from "../../../../utils/utils";
+import { EncryptedHeader } from "../encrypted_header";
 import { AbeEncryptionParameters } from "../encryption_parameters";
-import { EncryptedHeader, HybridEncryption } from "../hybrid_crypto";
+import { HybridEncryption } from "../interfaces/encryption";
 import { Metadata } from "../metadata";
