+++ src/crypto/abe/hybrid_crypto/gpsw/encryption.ts
@@ -2,8 +2,9 @@ import {
     webassembly_create_encryption_cache, webassembly_destroy_encryption_cache, webassembly_encrypt_hybrid_block, webassembly_encrypt_hybrid_header, webassembly_encrypt_hybrid_header_using_cache
 } from "abe_gpsw"
 import { logger } from "../../../../utils/logger"
+import { EncryptedHeader } from "../encrypted_header"
 import { AbeEncryptionParameters } from "../encryption_parameters"
-import { EncryptedHeader, HybridEncryption } from "../hybrid_crypto"
+import { HybridEncryption } from "../interfaces/encryption"
 import { Metadata } from "../metadata"
