+++ src/crypto/abe/hybrid_crypto/interfaces/encryption_parameters.ts
@@ -1,0 +1,1 @@
+export abstract class EncryptionParameters { }
