+++ src/demos/abe/cover_crypt/cover_crypt.ts
@@ -1,9 +1,9 @@
 import { webassembly_generate_master_keys, webassembly_generate_user_private_key } from "cover_crypt";
-import { CoverCryptHybridDecryption } from "../../crypto/abe/hybrid_crypto/cover_crypt/decryption";
-import { CoverCryptHybridEncryption } from "../../crypto/abe/hybrid_crypto/cover_crypt/encryption";
-import { AbeMasterKey } from "../../crypto/abe/keygen/keygen";
-import { Policy } from "../../crypto/abe/keygen/policy";
-import { fromBeBytes, hexDecode } from "../../utils/utils";
+import { CoverCryptHybridDecryption } from "../../../crypto/abe/hybrid_crypto/cover_crypt/decryption";
+import { CoverCryptHybridEncryption } from "../../../crypto/abe/hybrid_crypto/cover_crypt/encryption";
+import { AbeMasterKey } from "../../../crypto/abe/keygen/keygen";
+import { Policy } from "../../../crypto/abe/keygen/policy";
+import { fromBeBytes, hexDecode } from "../../../utils/utils";
 
 export function generateMasterKeys(policy: Policy): AbeMasterKey {
     const policyBytes: Uint8Array = policy.toJsonEncoded();
