+++ src/demos/abe/cover_crypt/demo_keys.ts
@@ -1,5 +1,5 @@
 
-import { hexDecode } from "../../../../utils/utils"
+import { hexDecode } from "../../../utils/utils"
 import { DemoKeys } from "../demo_hybrid_crypto"
