+++ src/demos/abe/gpsw/demo_keys.ts
@@ -1,7 +1,5 @@
-import { hexDecode } from "../../../../utils/utils"
+import { hexDecode } from "../../../utils/utils"
 import { DemoKeys } from "../demo_hybrid_crypto"
-
-
 
 export class GpswDemoKeys extends DemoKeys {
